modified authenticationPrincipal for more secure

+++ user-service/src/main/java/com/CNTTK18/user_service/controller/UserController.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.CNTTK18.user_service.data.Role;
 import com.CNTTK18.user_service.dto.request.AddressRequest;
 import com.CNTTK18.user_service.dto.request.Login;
 import com.CNTTK18.user_service.dto.request.Password;
@@ -26,11 +28,13 @@ import com.CNTTK18.user_service.dto.request.UserUpdateAfterLogin;
 import com.CNTTK18.user_service.dto.response.AddressResponse;
 import com.CNTTK18.user_service.dto.response.TokenResponse;
 import com.CNTTK18.user_service.dto.response.UserResponse;
+import com.CNTTK18.user_service.model.UserPrinciple;
 import com.CNTTK18.user_service.service.UserService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import jakarta.ws.rs.ForbiddenException;
 
 @RestController
 @RequestMapping("/api/users")
@@ -105,7 +109,12 @@ public class UserController {
     @Tag(name = "Put")
     @Operation(summary = "Update user")
     @PutMapping("/{id}")
-    public ResponseEntity<UserResponse> updateUser(@PathVariable String id, @RequestBody @Valid UserRequest updateUserDTO) {
+    public ResponseEntity<UserResponse> updateUser(@AuthenticationPrincipal UserPrinciple user, @PathVariable String id, 
+                                                    @RequestBody @Valid UserRequest updateUserDTO) {
+
+        if (user == null || (!user.getId().equals(id) && !user.getRole().equals(Role.ADMIN.toString()))) {
+            throw new ForbiddenException("You are not allowed to update this user.");
+        }
         UserResponse updatedUser = userService.updateUser(id, updateUserDTO);
         return ResponseEntity.ok(updatedUser);
     }
@@ -113,7 +122,10 @@ public class UserController {
     @Tag(name = "Delete")
     @Operation(summary = "Delete user")
     @DeleteMapping("/{id}")
-    public ResponseEntity<MessageResponse> deleteUser(@PathVariable String id) {
+    public ResponseEntity<MessageResponse> deleteUser(@AuthenticationPrincipal UserPrinciple user, @PathVariable String id) {
+        if (user == null || (!user.getId().equals(id) && !user.getRole().equals(Role.ADMIN.toString()))) {
+            throw new ForbiddenException("You are not allowed to update this user.");
+        }
         userService.deleteUserById(id);
         return ResponseEntity.ok(new MessageResponse("User deleted successfully"));
     }
